refactor: moved bluetooth scanner to infra
