--- app/src/main/java/jp/aoyama/mki/thermometer/view/bluetooth/scanner/BluetoothDeviceData.kt
@@ -1,10 +1,0 @@
-package jp.aoyama.mki.thermometer.view.bluetooth.scanner
-
-import android.bluetooth.BluetoothDevice
-import java.util.*
-
-data class BluetoothDeviceData(
-    val device: BluetoothDevice,
-    val rssi: Int,
-    val foundAt: Calendar,
-)
